Remove unnecessary code from DocumentMarkerController.
RepaintMarkers was only used by spelling marker code and find-in-page
code. Remove it because ...

The spelling code already calls RemoveSpellingMarkersUnderWords which
invalidates all the nodes that change status.

The calls for search text markers are only made by code that sets
markers as painted or not. But that code is redundant: we always
paint find-in-page markers when they are added, and invalidate when
markers are removed. Remove the method in question and all the
internal code that allows tests to do something they now don't
need to do.

Bug: 368113219
Change-Id: I6cb5f437be7eb33de331d66f6017d5b9616cbc0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5884274
Reviewed-by: Philip Rogers <pdr@chromium.org>
Auto-Submit: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359103}
NOKEYCHECK=True
GitOrigin-RevId: f3fae8c3c1c805156cf7a5a8124db274bdd7c355