Refactor selectlist tests [1/4]
This is a set of refactors that:
 1. re-points existing WPTs and web_tests from <selectlist> to
    <select style="appearance:base".
 2. renames tests from `-selectlist-` to `-select-`

These are done in batches, to make the review task easier. One CL
will re-point tests (without renaming them) and the next will do
the rename.

These are all in service of removing the <selectlist> prototype.

Bug: 335458492
Change-Id: Ie8dfdc37f1c3a6323d526ca1979d315c111d93f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857249
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359127}
NOKEYCHECK=True
GitOrigin-RevId: e62f9f55634a26228a93aa4e0b11af8a376dd9de