Refactor selectlist tests [2/4]
This is a set of refactors that:
 1. re-points existing WPTs and web_tests from <selectlist> to
    <select style="appearance:base".
 2. renames tests from `-selectlist-` to `-select-`

These are done in batches, to make the review task easier. One CL
will re-point tests (without renaming them) and the next will do
the rename.

These are all in service of removing the <selectlist> prototype.

Bug: 335458492
Change-Id: I4315b121688e0becdf7edc96287d6cd805439d24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857250
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359130}
NOKEYCHECK=True
GitOrigin-RevId: e67eb82ad0a0a6c52190db7571ae55097975e95d