Refactor selectlist tests [3/4]
This is a set of refactors that:
 1. re-points existing WPTs and web_tests from <selectlist> to
    <select style="appearance:base".
 2. renames tests from `-selectlist-` to `-select-`

These are done in batches, to make the review task easier. One CL
will re-point tests (without renaming them) and the next will do
the rename.

These are all in service of removing the <selectlist> prototype.

Bug: 335458492
Change-Id: Iac8e531e0c3354e2637350218e6e4e43002f6747
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857251
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359131}
NOKEYCHECK=True
GitOrigin-RevId: f3f413df4cf8d8b6a8e7d3b98616419627d4dcce