[carousel] Layout support for ::column::scroll-marker.
Modify Element::ScrollIntoViewNoVisualUpdate() to allow pseudo elements,
but locate the originating element when needed.  We need to call
BoundingBoxForScrollIntoView() on the pseudo element for ::column scroll
markers, since the originating element is the multicol container itself.
But we want individual columns.

Some extra care was required for fieldsets. This is why
EnclosingDOMNode() was added. If the element is a fieldset, the
scrollable container and/or multicol container is an anonymous fieldset
content child of the fieldset.

Also need to make sure that ::column scroll markers aren't focusable, as
that would trigger style recalc when focused, and we don't handle that
yet.

Bug: 365680822
Change-Id: Ia735a21ce2f4a236ce6addcd8653fae99e59e5c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5872577
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Auto-Submit: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359224}
NOKEYCHECK=True
GitOrigin-RevId: 62e3df7878c6eb44c25d761a0bc1a5738ba8ac4b