NavPredictor: Send "LeftViewport" notification after anchor unobserved
Sends a "left viewport" notification to the browser after we
deprioritize observing an anchor with IntersectionObserver. This is
to make sure we don't record any stale viewport data on the browser
side if the deprioritized anchor was in the viewport when it was
unobserved.

Bug: 331670939
Change-Id: I6c5d9837156e86fdc8e5ebe31d300cdbfcec4b72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5874366
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359342}
NOKEYCHECK=True
GitOrigin-RevId: b06ba493ccdac63e5215f4dfd2b58d9c82367b76