[webcodecs] Remove SIFormatType::Legacy check from CopyToFormat
CopyToFormat checks for RequiresExternalSampler and returns early if
so. Accordingly it should never reach the conditional if the format is
multiplanar and thus it should be single planar. In such cases, format
type can never be kSharedImageFormatExternalSampler and NumPlanes
must be 1 and the condition with Legacy can never be true. This
removes the check and simplifies logic.

Bug: 343011436
Change-Id: I968af21f5ee50f72cd99260f78327c33c27c5a89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5881868
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359370}
NOKEYCHECK=True
GitOrigin-RevId: 473b5ed58f95d142232dda64dfd8dcb4b65cec74