Replace StreamPromiseResolver with ScriptPromiseResolver
Their behavior is close enough that it doesn't really makes sense to
have two separate implementation of promise resolution.

The main substantive difference is that StreamPromiseResolver
doesn't appear to check whether the ExecutionContext is detached
before resolving. No tests appear to require resolving after detach,
but StreamPromiseResolver doesn't allow it, so it is a change.

ScriptPromiseResolver is typed, so although the promises implemented
by StreamPromiseResolver don't have associated IDL to determine the
resolve type, I've imputed it based on what we actually resolve with
in the implementation.

Change-Id: I91d6c7d570904a302c3125b6edc0f097be61d538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5871252
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359524}
NOKEYCHECK=True
GitOrigin-RevId: 01c6cb855ca0d550bf0184f97164a3e6fbadb9ee