[inspector] Fix accessing disposed V8 session in page agent
Scripts running as part of Page.evaluateScriptOnNewDocument can pause
the page. During a pause we can detach the DevTools session, but the
page agent is still in the middle of running the
"DidCreateMainWorldContext" probe. This means any additional
Page.evaluateScriptOnNewDocument scripts would attempt to eval on
a detached V8 session.

This CL fixes this by overriding InspectorBaseAgent::Dispose in the
page agent and resetting `v8_session_` to a nullptr which we can
check for before evaling more scripts.

This check is only necessary for page agent methods that execute
more than one JS script as for all the others we wouldn't call
the probes on a disposed agent in the first place.

R=caseq@chromium.org, dsv@chromium.org

Fixed: 368672129
Change-Id: I4c3361c8116a64343206da991e503aaa6bd917f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5886170
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1359730}
NOKEYCHECK=True
GitOrigin-RevId: 622d2fd964161261c4a5afeaad7df547bb04e0aa