Make AudioContext.setSinkId respond to permission changes
This is a combination / extension of:
https://crrev.com/c/5850652
https://crrev.com/c/5841286

Currently setSinkId will not update the cached device list when the
permission changes.  This may result in allowing setSinkId when it
should be blocked, or vice-versa.

Instead, update the device list when permission changes and block
resolving setSinkId calls until the enumeration is complete.

Bug: 364232288
Bug: 328109813
Change-Id: I32bb0af37c30d38569615f803a95fc0c64de5908
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5876827
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1360302}
NOKEYCHECK=True
GitOrigin-RevId: 84a4e92d5b48c73e2145fa9e2cf655cc232b1928