Support for @page counters.
We need to create a deep clone of CountersAttachmentContext for each
page, since we may have to discard the page layout in some cases and
retry. Which means that we also need to roll back what was done to the
counters.

Page counters are rather special, when reset, they will obscure any
counters in the document (rather than establishing a new scope). This
even happens when incrementing page counters in some cases. Update
CountersAttachmentContext to handle this.

And, as always for block fragmentation, the out-of-flow layout code
needs a little special-treatment. If out-of-flow elements causes
additional pages to be created, we need to go back and reprocess all
page counters.

Bug: 40529222
Change-Id: I7c9c80114191ae760eab662ea7fd345237498945
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5887168
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1360458}
NOKEYCHECK=True
GitOrigin-RevId: cab22cc1c5e65b250da8c129f4c47731592947d3