Add 'webgpu' WebXR feature
The 'webgpu' feature enables the session to use WebGPU layers and
prevents the use of WebGL layers. It also changes the projection
matrix calculations to match WebGPU's clip space.

Worth noting that actually attempting to create a layer with this feature right now will result in the session being rejected since the backends don't indicate that it's supported. This is just an incremental step towards the feature support.

Bug: 359418629
Change-Id: If22e6c8a6b9c6d03b935601e13d9a2972099c86e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5868084
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361344}
NOKEYCHECK=True
GitOrigin-RevId: 9475850f43214842d01be9893350ef9ae49afcfb