Introduce HTMLAnchorElementBase
This CL introduces HTMLAnchorElementBase as an abstraction to
represent both <a> and <area> (this is used particularly in
speculation rules code), and allow for safer type casting. See bug for
rationale.

Note: There is no behaviour change in this patch, it is a pure
refactor. There are comments added in places where an incorrect type
cast is being done, which will be fixed in follow-ups (if necessary).

Bug: 369219144
Change-Id: I16d2c9bbacfdf8342e19deea64f68c87e18bfbb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5890369
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361822}
NOKEYCHECK=True
GitOrigin-RevId: 54a24ba29a26b8ffebe978903de46850b8fea58b