createImageBitmap: Respect colorSpaceConversion for video
Plumb the parameter to reinterpret the video as being in sRGB from
- blink::ImageBitmap::ImageBitmap (read from options)
- blink::HTMLVideoElement::CreateStaticBitmapImage
- blink::CreateImageFromVideoFrame
- media::PaintCanvasVideoRenderer::Paint (via new parameter)

Add a WPT test for this. For the WPT test, add a Display P3 mp4 and
webm test video. (There exists a Rec2020 test video, but it is 10-bit,
which has bugs on some platforms, and uses the Rec709 transfer function,
which has different interpretations on all browsers).

Bug: 362431889
Change-Id: I3c12eb720b927a8e6466ab35ae29a044342df123
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5878044
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361907}
NOKEYCHECK=True
GitOrigin-RevId: fea386a9971233d2539659052569c90ed9c234b5