Update GPURequestAdapterOptions.featureLevel, make it experimental
This changed from `any` to `DOMString` in the spec.

Also gate `featureLevel` behind `WebGPUExperimentalFeatures` until we do
the Intent to Ship. It is web-observable by whether we validate its
value or not (and whether we access the property).

There is no change to the handling; we still don't allow this to be set
to any value (per spec), and return `null` if it is.

Bug: 366151404
Change-Id: I24fb346e993a619ec167f0c8c21861b46e849ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5896612
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Auto-Submit: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362070}
NOKEYCHECK=True
GitOrigin-RevId: 4f28d0070e76f7b3871b43fd87cef5d1a4bd2097