Enable flaky endorser on Wpt importer
To get a sense how often a flaky test will be imported.

If an imported CL is blocked due to this, the gardener could add
'Validate-Test-Flakiness: skip' to the CL footer, and created a
bug on the flaky tests for test owners to fix.

We should revert this change if this happens too often, and implement
the ideas in http://shortn/_X2rMgQLT7h

Bug: 347713900
Change-Id: Id2ea143663598574fcc1afe6d92ee07de4f28828
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900511
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1362074}
NOKEYCHECK=True
GitOrigin-RevId: a21262d442e1653c8826165ffb9373b0a62c34a0