Build Crubit's rs_bindings_from_cc when building the Rust toolchain
We use the new cargo-based build for Crubit. The build is allowed to
fail still, without causing the package to fail, since we need to wait
until we can rely on the cargo-based Crubit build to remain green and
working with the latest Clang and Rust revisions.

Bug: 40226863
Change-Id: I43134d740729e8d88dd2448a0adf3fbef6179808
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5888109
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362506}
NOKEYCHECK=True
GitOrigin-RevId: 9a318f60c017464420bb99359a294b5007e13664