Reland: createImageBitmap: Add correct parameter computation in parallel
Landed in crrev.com/1361920, reverted in crrev.com/1361927 due to
mid-air collision.

The createImageBitmap function specifies the source rect to create
the bitmap from. This rect is in coordinates after applying the
source's orientation.

If no source rect is specified, then we should set it to the full
source. There is a bug that when we do this, we use the source's
size before applying the source's orientation.

Tentatively and with great terror, start moving towards fixing this.

The first issue is that we do not know the source's orientation in
ParseOptions. Re-arrange the calls to ParseOptions so that the orientation is present when it is needed.

Do not actually fix the computation, because there are other bugs that
sometimes cancel it out. Instead, add correct values in parallel
(source_orientation, source_size, source_rect, dest_rect).

The next patch will create a new and more compact version of the
code that currently applies the requested transformations.

Bug: 40773069, 362431889
Change-Id: I151426026328a48ce7bd9db1ae43be3875681c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5899179
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362570}
NOKEYCHECK=True
GitOrigin-RevId: 4af167ab062f96007e01725c89fa4ea7c5a18c00