Add Mutation Observer for customizable <select> and console message
Introduction of the mutation observer for <select> to detect whenever there's a change in its descendants and addition of a console message for every descendant that's not conforming to the content model (left TODOs that will be addressed in future CLs). Tests were added to validate behavior. Test expectations were updated to add the new console warning when appropriate.
Follow-ups: report issues to the DevTools' Issue Panel & look at the tree structure to decide whether a descendant is valid or not.

Bug: 347890366
Change-Id: I9d953d68360d388e8227cfbdf380b5933a0dcdb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5856019
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362734}
NOKEYCHECK=True
GitOrigin-RevId: f6e53c879e1dc17b446006429680b884739d1bab