createImageBitmap: Add fused software path
Add the function ApplyTransformsFromOptionsSoftware, which performs
the same operations as ApplyTransformsFromOptionsWithBlit, except
is directly manipulates SkPixmaps.

Add the helper function FlipSkPixmapInPlace which is an almost-copy-
paste of the code in FlipImageVertically. That function will be
deleted.

Move more ImageBitmap paths to the new path.

Bug: 40773069, 362431889
Change-Id: I53c10521ab8cb91429f8b92e01b3b9b1c6501385
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5893617
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362744}
NOKEYCHECK=True
GitOrigin-RevId: bccce7a992ba003f8d6d5f4dd71b72b399fb8ccc