Move StyleSheetHandler to separate file, as InspectorCSSParserObserver
In order to work towards a fix for Issue 363985597, it will be necessary
to write unit tests against StyleSheetHandler. This is quite hard to
do while it's an implementation detail of InspectorStylesheet, hence
this CL moves it to a separate file.

There is no behavior change.

Bug: 363985597
Change-Id: I4c6897db06faa23b1f15f05cd8206d2ed9d38de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904154
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362983}
NOKEYCHECK=True
GitOrigin-RevId: c861d24f0ccac150f26048f2cd9db2ca60095847