Remove unused random-access blob method from Blink.
The ReadRange method in the mojom is left in place because it's
used by some browser code (for example, CrOS FileManager). It's
strange to have a Blink API contain code that is not used by
or in service of Blink, but refactoring that is left as a TODO.
(It's also worth noting that all callers of the mojom only want
to read the /first/ N bytes, i.e. the offset is always 0.)

Bug: none
Change-Id: I54b619c753521e1c06fd6a6c95e57d4bca7cba35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5878416
Reviewed-by: Brad Triebwasser <btriebw@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363107}
NOKEYCHECK=True
GitOrigin-RevId: 88b42db2e98230779171c9489443049e512945b3