Create JS module with correct world
When a JS module is created, the process is asynchronous, so we need
to make sure that the correct world is used. In the case of sandboxed
frames and injected extension scripts, this doesn't currently happen,
since the extension's content script needs to run in an isolated world
that is not the main world, and calls to
ExecutionContext::CanExecuteScript need to know that an isolated
world does exist during the module creation when the module is being
created by a extension.

Bug: 40564567
Change-Id: I30fd6e2b1fcce7a95dc5bc6458fca796c00a687a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860258
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363356}
NOKEYCHECK=True
GitOrigin-RevId: b15ccc1ec096516cef58791d9ded02fd74e02b7f