Revise the behavior of wpt-tests when the object element is not displayed
According to the spec[1], the `<object>` element that is not rendered
will not fetch request.
> if the element is not being rendered, then jump to the step below
> labeled fallback.
> Fallback: The object element represents the element's children.
In the `object-in-object-fallback-2.html` test case, all `<object>`
elements are not being displayed, so no elements should be loaded.
Remove `display: none` in this test case to test the `<object>`
element nesting inside the `<object>` element.

Revise the `object-in-display-none-load-event.html` test case to
ensure that the `object` element is not loaded.

Bug: 40781343
Change-Id: I190c937dbe643f5d1502529f8a530687ed8fcb9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904314
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Traian Captan <tcaptan@chromium.org>
Commit-Queue: Peng Zhou <zhoupeng.1996@bytedance.com>
Cr-Commit-Position: refs/heads/main@{#1363461}
NOKEYCHECK=True
GitOrigin-RevId: 2060de7023ec247e7b8d739532437761c8285792