Expand ServiceWorkerRouterRaceSource to support other races
In crbug.com/370844790, we intend to add support on racing between
network and cache storage.  For that case, ServiceWorkerRouterRaceSource
should have the way to describe sources that race.

This is an initial change to expand the current system to support
multiple sources race.

Bug: 370844790
Change-Id: I7414eaf5b6666bcb91a27427db65919934718160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904958
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363508}
NOKEYCHECK=True
GitOrigin-RevId: 0049ab4b8a7db5cfe21ec6b2fa5c4b1cb4fbd325