[Blink] Make CanvasHibernationHandler tests more self-contained
The CanvasHibernationHandler tests currently rely on Canvas2DLayerBridge
logic for triggering and ending hibernation. This is problematic as we
are in the process of moving this logic from Canvas2DLayerBridge to
CanvasRenderingContext2D, and it will not be possible to instantiate
CanvasRenderingContext2D in the context of these tests.

The logic that the tests rely on is simply that hibernation is triggered
and ended in response to page visibility changing, i.e. [1] and [2].
This CL changes the tests to inline this logic directly rather than
calling into Canvas2DLayerBridge. This makes the tests more
self-contained. Coverage of the Canvas2DLayerBridge logic is maintained
by existing Canvas2DLayerBridge and CanvasRenderingContext2D tests of
hibernation.  We will be porting the Canvas2DLayerBridge tests to be on
CanvasRenderingContext2D, so this coverage will be preserved.

A few notes:

* I changed the flow for triggering hibernation to occur synchronously.
  This is in effect what the code is doing now, since after calling
  into Canvas2DLayerBridge it immediately runs all idle tasks to ensure
  that hibernation is triggered before returning to the caller.
* One test verifies that an invocation of GetImage() results in a given
  metric being recorded. This method is no longer automatically invoked
  in the flow of ending hibernation (since I stripped that down to just
  calling Clear() on the hibernation handler), so I invoke GetImage()
  explicitly before doing the verification in question.
* One test implicitly relies on the fact that invoking
  GetOrCreateResourceProvider() on Canvas2DLayerBridge ends hibernation
  if the bridge has previously dropped the resource provider as part of
  hibernation beginning. As we are no longer going through the bridge
  for the flow of hibernation, this implicit reliance no longer holds.
  We instead just explicitly end hibernation at that point in the test.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_2d_layer_bridge.cc;l=128-142?q=canvas_2d_layer_brid&ss=chromium
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_2d_layer_bridge.cc;l=224;drc=f522344e45882da4c7f7cb1b3a0a7bd747d654bb?q=canvas_2d_layer_brid&ss=chromium

Bug: 40280152
Change-Id: I239fe02da057c298c27024c508099339d2bb42ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5899738
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363596}
NOKEYCHECK=True
GitOrigin-RevId: 9305e44079f8b75be71a7e980d4d71cabeb90d39