Remove include of flow_event_utils.h from latency_info.h
This include is very costly according to
https://commondatastorage.googleapis.com/chromium-browser-clang/include-analysis.html#view=edges

This is not used, other than for indirect includes of other files. We
replace it with a more targeted include and a forward declaration for
what the file actually references, and also fix an IWYU issue.

Bug: 40318405
Change-Id: I0897ff186732c5048cfc9bb77f250715fc209fc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5887793
Auto-Submit: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Robert Flack <flackr@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363748}
NOKEYCHECK=True
GitOrigin-RevId: e8ecbd4efd867c4cfa4547878cbf73374cdcfce3