[flex] Don't apply transferred min/max sizes.
https://chromium.googlesource.com/chromium/src/+/f357fe9f912bd1723fb73fdb693c6825dae23390

... introduced a change where we applied the transferred min/max sizes
to the inline min/max sizes.

This was unintentional. The spec is unclear on this situation (after
checking with the spec author), so filed an issue here:
https://github.com/w3c/csswg-drafts/issues/10997

We may change back again, but at least this brings us in-line with
other browsers for the moment.

Fixed: 368562235
Change-Id: I6f353e7d2e1ef0e1077af2b3b76c5f42632cc6b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5907910
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1363796}
NOKEYCHECK=True
GitOrigin-RevId: e66cbb02eb477bab76e3c2616f4e45e8143e2e2f