Consider a clip-path url() reference an empty clip while it's loading
While an external clip path is loading, we'll currently render as if no
clip path is applied. Not applying a clip is what the spec prescribes
for invalid references. While the resource is loading though, we
shouldn't consider it invalid. Instead let an empty Path represent the
external reference while it is loading.

Fixed: 371112520
Change-Id: Icbdf1d6bf6f981f53c139df22d320b36afa1aab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5905324
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1364171}
NOKEYCHECK=True
GitOrigin-RevId: d2afc9045478de88b90555868743724318dd681f