Add ScriptState null check in AnimationFrameTimingMonitor
There is a crash report which shows that ToScriptStateForMainWorld is
returning null and hitting a check after null it passed into
PopScriptEntryPoint. This patch fixes this by checking whether the
ScriptState is null.

It looks like this case can happen when the frame is attached and
provisional, which causes ToV8ContextEvenIfDetached to return an empty
context which in return makes ToScriptStateForMainWorld to return null.

Fixed: 360742350
Change-Id: Ib5f36c4ed6413b26e41571262715559dfe0eeb21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5907916
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Yoav Weiss (@Shopify) <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1364306}
NOKEYCHECK=True
GitOrigin-RevId: 0e12358c50f32d6312f010aef48660ebc5b10938